In disk_capacity_notwindows.go , cast stat.Blocks and stat.Bavail explicitly to uint64 before
multiplication:
go totalBytes := uint64(stat.Blocks) * uint64(stat.Bsize)
freeBytes := uint64(stat.Bavail) * uint64(stat.Bsize)

This resolves the platform-dependent mismatched types compiler error (int64 vs uint64) on unix-based builds.

+++ internal/api/disk_capacity_notwindows.go
@@ -11,7 +11,7 @@ func getDiskCapacity(path string) (total uint64, free uint64, err error) {
 		return 0, 0, err
 	}
 
-	totalBytes := stat.Blocks * uint64(stat.Bsize) //nolint:unconvert
-	freeBytes := stat.Bavail * uint64(stat.Bsize)  //nolint:unconvert
+	totalBytes := uint64(stat.Blocks) * uint64(stat.Bsize) //nolint:unconvert
+	freeBytes := uint64(stat.Bavail) * uint64(stat.Bsize)  //nolint:unconvert
 	return totalBytes, freeBytes, nil
 }
